Shorten B117 plots to end mid May. Make sure all x lims of the plots are the same.
